+++ Frends.JSON.Handlebars/Frends.JSON.Handlebars/Definitions/Options.cs
@@ -1,0 +1,25 @@
+﻿using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
+
+namespace Frends.JSON.Handlebars.Definitions;
+
+/// <summary>
+/// Options parameters.
+/// </summary>
+public class Options
+{
+    /// <summary>
+    /// Whether to throw an exception on failure or return a Result with Success = false.
+    /// </summary>
+    /// <example>true</example>
+    [DefaultValue(true)]
+    public bool ThrowErrorOnFailure { get; set; } = true;
+
+    /// <summary>
+    /// Custom error message used when an error occurs.
+    /// </summary>
+    /// <example></example>
+    [DisplayFormat(DataFormatString = "Text")]
+    [DefaultValue("")]
+    public string ErrorMessageOnFailure { get; set; } = string.Empty;
+}